requirements.txt
jinja2==3.1.6
pyyaml==6.0.3
paramiko
scp
python-dotenv
numpy
scipy